refactors single page express app to express routes
